Fix entity, controller, bug lombok và chạy app

+++ sbquanan/src/main/java/com/example/sbquanan/controller/CaLamViecController.java
@@ -15,19 +15,19 @@ public class CaLamViecController {
     @GetMapping public List<CaLamViec> getAll() { return repository.findAll(); }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CaLamViec> getById(@PathVariable Long id) {
+    public ResponseEntity<CaLamViec> getById(@PathVariable Integer id) {
         return repository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping public CaLamViec create(@RequestBody CaLamViec e) { return repository.save(e); }
 
     @PutMapping("/{id}")
-    public ResponseEntity<CaLamViec> update(@PathVariable Long id, @RequestBody CaLamViec updated) {
+    public ResponseEntity<CaLamViec> update(@PathVariable Integer id, @RequestBody CaLamViec updated) {
         return repository.findById(id).map(e -> { updated.setCaID(e.getCaID()); return ResponseEntity.ok(repository.save(updated)); }).orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         if (!repository.existsById(id)) return ResponseEntity.notFound().build();
         repository.deleteById(id); return ResponseEntity.noContent().build();
     }
