+++ sbquanan/src/main/java/com/example/sbquanan/controller/KetCaController.java
@@ -15,19 +15,19 @@ public class KetCaController {
     @GetMapping public List<KetCa> getAll() { return repository.findAll(); }
 
     @GetMapping("/{id}")
-    public ResponseEntity<KetCa> getById(@PathVariable Long id) {
+    public ResponseEntity<KetCa> getById(@PathVariable Integer id) {
         return repository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping public KetCa create(@RequestBody KetCa e) { return repository.save(e); }
 
     @PutMapping("/{id}")
-    public ResponseEntity<KetCa> update(@PathVariable Long id, @RequestBody KetCa updated) {
+    public ResponseEntity<KetCa> update(@PathVariable Integer id, @RequestBody KetCa updated) {
         return repository.findById(id).map(e -> { updated.setKetCaID(e.getKetCaID()); return ResponseEntity.ok(repository.save(updated)); }).orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         if (!repository.existsById(id)) return ResponseEntity.notFound().build();
         repository.deleteById(id); return ResponseEntity.noContent().build();
     }
