+++ sbquanan/src/main/java/com/example/sbquanan/controller/KhuyenMaiController.java
@@ -15,19 +15,19 @@ public class KhuyenMaiController {
     @GetMapping public List<KhuyenMai> getAll() { return repository.findAll(); }
 
     @GetMapping("/{id}")
-    public ResponseEntity<KhuyenMai> getById(@PathVariable Long id) {
+    public ResponseEntity<KhuyenMai> getById(@PathVariable Integer id) {
         return repository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping public KhuyenMai create(@RequestBody KhuyenMai e) { return repository.save(e); }
 
     @PutMapping("/{id}")
-    public ResponseEntity<KhuyenMai> update(@PathVariable Long id, @RequestBody KhuyenMai updated) {
+    public ResponseEntity<KhuyenMai> update(@PathVariable Integer id, @RequestBody KhuyenMai updated) {
         return repository.findById(id).map(e -> { updated.setKhuyenMaiID(e.getKhuyenMaiID()); return ResponseEntity.ok(repository.save(updated)); }).orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         if (!repository.existsById(id)) return ResponseEntity.notFound().build();
         repository.deleteById(id); return ResponseEntity.noContent().build();
     }
