+++ sbquanan/src/main/java/com/example/sbquanan/controller/MenuController.java
@@ -15,19 +15,19 @@ public class MenuController {
     @GetMapping public List<Menu> getAll() { return repository.findAll(); }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Menu> getById(@PathVariable Long id) {
+    public ResponseEntity<Menu> getById(@PathVariable Integer id) {
         return repository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping public Menu create(@RequestBody Menu e) { return repository.save(e); }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Menu> update(@PathVariable Long id, @RequestBody Menu updated) {
+    public ResponseEntity<Menu> update(@PathVariable Integer id, @RequestBody Menu updated) {
         return repository.findById(id).map(e -> { updated.setMenuID(e.getMenuID()); return ResponseEntity.ok(repository.save(updated)); }).orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         if (!repository.existsById(id)) return ResponseEntity.notFound().build();
         repository.deleteById(id); return ResponseEntity.noContent().build();
     }
