+++ sbquanan/src/main/java/com/example/sbquanan/controller/MonAnController.java
@@ -15,19 +15,19 @@ public class MonAnController {
     @GetMapping public List<MonAn> getAll() { return repository.findAll(); }
 
     @GetMapping("/{id}")
-    public ResponseEntity<MonAn> getById(@PathVariable Long id) {
+    public ResponseEntity<MonAn> getById(@PathVariable Integer id) {
         return repository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     @GetMapping("/menu/{menuId}")
-    public List<MonAn> getByMenu(@PathVariable Long menuId) {
+    public List<MonAn> getByMenu(@PathVariable Integer menuId) {
         return repository.findByMenu_MenuID(menuId);
     }
 
     @PostMapping public MonAn create(@RequestBody MonAn e) { return repository.save(e); }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         if (!repository.existsById(id)) return ResponseEntity.notFound().build();
         repository.deleteById(id); return ResponseEntity.noContent().build();
     }
