+++ sbquanan/src/main/java/com/example/sbquanan/controller/NhanVienController.java
@@ -15,19 +15,19 @@ public class NhanVienController {
     @GetMapping public List<NhanVien> getAll() { return repository.findAll(); }
 
     @GetMapping("/{id}")
-    public ResponseEntity<NhanVien> getById(@PathVariable Long id) {
+    public ResponseEntity<NhanVien> getById(@PathVariable Integer id) {
         return repository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping public NhanVien create(@RequestBody NhanVien e) { return repository.save(e); }
 
     @PutMapping("/{id}")
-    public ResponseEntity<NhanVien> update(@PathVariable Long id, @RequestBody NhanVien updated) {
+    public ResponseEntity<NhanVien> update(@PathVariable Integer id, @RequestBody NhanVien updated) {
         return repository.findById(id).map(e -> { updated.setId(e.getId()); return ResponseEntity.ok(repository.save(updated)); }).orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         if (!repository.existsById(id)) return ResponseEntity.notFound().build();
         repository.deleteById(id); return ResponseEntity.noContent().build();
     }
