+++ sbquanan/src/main/java/com/example/sbquanan/repository/CaLamViecRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CaLamViecRepository extends JpaRepository<CaLamViec, Long> {}
+public interface CaLamViecRepository extends JpaRepository<CaLamViec, Integer> {}
