+++ sbquanan/src/main/java/com/example/sbquanan/repository/ChiTietDonHangRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChiTietDonHangRepository extends JpaRepository<ChiTietDonHang, Long> {}
+public interface ChiTietDonHangRepository extends JpaRepository<ChiTietDonHang, Integer> {}
