+++ sbquanan/src/main/java/com/example/sbquanan/repository/DonHangRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DonHangRepository extends JpaRepository<DonHang, Long> {}
+public interface DonHangRepository extends JpaRepository<DonHang, Integer> {}
