+++ sbquanan/src/main/java/com/example/sbquanan/repository/KetCaRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KetCaRepository extends JpaRepository<KetCa, Long> {}
+public interface KetCaRepository extends JpaRepository<KetCa, Integer> {}
