+++ sbquanan/src/main/java/com/example/sbquanan/repository/KhachHangRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KhachHangRepository extends JpaRepository<KhachHang, Long> {}
+public interface KhachHangRepository extends JpaRepository<KhachHang, Integer> {}
