+++ sbquanan/src/main/java/com/example/sbquanan/repository/KhuyenMaiRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KhuyenMaiRepository extends JpaRepository<KhuyenMai, Long> {}
+public interface KhuyenMaiRepository extends JpaRepository<KhuyenMai, Integer> {}
