+++ sbquanan/src/main/java/com/example/sbquanan/repository/ThanhToanRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ThanhToanRepository extends JpaRepository<ThanhToan, Long> {}
+public interface ThanhToanRepository extends JpaRepository<ThanhToan, Integer> {}
